requirements.txt
tornado
requests
pyjwt
gitpython
there
mock
cryptography
friendlyautopep8
mock == 2.0.0
